requirements.txt
pandas
numpy
scikit-learn
matplotlib
plotly
streamlit
folium
streamlit-folium
mlflow
umap-learn
